backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "arkiv-sdk>=1.0.0a8",
    "python-dotenv>=1.2.1",
]
